提交
优化提交

+++ TodoMacRepertorie/TodoMacRepertorie/TDTask(任务列表)/model/TDGetCurrentVersionModel.swift
@@ -9,5 +9,5 @@ import Foundation
 
 /// 服务器返回的最大版本号响应
 struct TDGetCurrentVersionModel: Codable {
-    let maxVersion: Int
+    let maxVersion: Int64
 }
